refactor: ユーザー詳細画面のファイル移動
