+++ src/pages/user/detail/[userId]/index.tsx
@@ -1,0 +1,5 @@
+import { UserDetail } from '../../../../presentation/components/UserDetail/UserDetail';
+
+export default function UserDetailPage() {
+  return <UserDetail />;
+}
